Refactor daemon module and enhance database API structure
This commit reorganizes the `daemon.rs` file by splitting it into multiple submodules, improving code organization and maintainability. It also introduces new database-related modules for ClickHouse, DuckDB, and PostgreSQL, along with a health check mechanism for backend services. The changes streamline the database initialization process and enhance the overall structure of the API, making it easier to manage and extend in the future.
