+++ bitloops/src/host/devql/query/dsl_compiler/types.rs
@@ -1,0 +1,13 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum GraphqlCompileMode {
+    Global,
+    Slim,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(super) enum RegisteredStageKind<'a> {
+    Tests(&'a super::RegisteredStageCall),
+    Coverage,
+    Knowledge(&'a super::RegisteredStageCall),
+    Extension(&'a super::RegisteredStageCall),
+}
