+++ bitloops/src/lib.rs
@@ -5,8 +5,8 @@ pub mod capability_packs;
 pub mod cli;
 pub mod config;
 pub mod daemon;
-pub(crate) mod devql_transport;
 pub(crate) mod devql_timing;
+pub(crate) mod devql_transport;
 pub mod git;
 pub mod graphql;
 pub mod host;
